python-sdk/pyproject.toml
[tool.poetry]
name = "hatchet-sdk"
version = "0.21.1"
description = ""
authors = ["Alexander Belanger <alexander@hatchet.run>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.60.0"
python-dotenv = "^1.0.0"
protobuf = "^4.25.2"
pyyaml = "^6.0.1"
grpcio-tools = "^1.60.0"
loguru = "^0.7.2"
pydantic = "^2.6.3"
python-dateutil = "^2.9.0.post0"
urllib3 = "^2.2.1"
aiostream = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_third_party = [
  "grpcio",
  "grpcio_tools",
  "loguru",
  "protobuf",
  "pydantic",
  "python_dotenv",
  "python_dateutil",
  "pyyaml",
  "urllib3"
]
